Feat: Add a feature for input validation
Add a feature for input validation to check the filled/selected status of the compulsory fields in the form.

Simply, add a feature to verify that desired form field is filled or not
when submit button is pressed. If compulsory fields are filled then no
issue else error will display

Signed-off-by: @surajgirioffl