Update: Update remote database server
Update remote database server and set to a database of same server on
which application and web is executing.

Because the application server (pythonanywhere) doesn't support any
external database server connection in free acccount. We also can't
access the database from any outer services in free account.
So, database of pythonanywhere (application and web server) is added to
avoid the error 'Unable to establish connection with database'.

Signed-off-by: @surajgirioffl