Feat: Add feat. to generate filename
Add functionality to generate filename based on the request credentials
sent by user to the server.

If user sent invalid credentials then filename generated will be as per
that invalid credentials but there will no option to download pdf if
credentials are not valid or server status code is other than 200.

Signed-off-by: @surajgirioffl