Add: Add leading `RESULT OF ` text in name of PDF
Add leading 'RESULT OF ' text in the name of PDF while saving if result is
available (means status code received from the server is 200).

Also update leading text in name of PDF if result is not available (By
the way, button will not visible to save for normal user in this case :smile:).

Signed-off-by: @surajgirioffl