Fix: Fix import error
Fix import error when executing the application (app.py).

Means if same script (this script) is executing then 'import bca_db as bca' working perfectly
but if we are importing the same in app.py then it raises 'import error'. So, for
that case we update the import statement from previous to 'from . import
bca_db as bca'.

Signed-off-by: @surajgirioffl