Rename: Rename a method
Rename the method '__setSessionScopedAttributes'  to
'setSessionScopedAttributes' because main() is calling this method which
is not part of the class 'Files' then why do we make it to seem like
private.

This method is not used by any other method of the class (instance). So,
no need to use leading dunder to make it seems like a private member
function.

Signed-off-by: @surajgirioffl