Add: Add class `phone-only` in view tb header and data
Add the class 'phone-only' in the view table-header and table-data to
make view attribute (table-header) and table-data visible only on mobile
devices (phones).

Because on the phone when the user clicks on download then the file
starts downloading instantly and user is unable to view the docs before
downloading while in pc/laptop when the user clicks on the download button
then the file is opened in a new tab and the user can view the docs and
then download the file.

Feature-suggested-by: @abhiabhishekoffl
Signed-off-by: @surajgirioffl